Added code to enabled OLED_motion, masks DHT sensors read errors with previous read values, reduced output Significant Figures so graphing is more accurate, and changed the min config so the serial output on the 8266 works on the IDE.
To enable the OLED_motion option I added an if statement in the main loop so it writes the display if its not enabled, or clears it if it is. Then added a write function to Motion section that write the display if its enabled, or just the M if not.

Changed code to report previous read good values from DHT on sensor read errors. The quality of these sensors very, and the temperatures should not be changing that much with these read frequencies. So I added variable to store the previous read values in. If the read results in an error it outputs the previous read values, otherwise it saves the new value for use in the next read.

I reduced the output significant figures from the DHT on all outputs, as the tolerance of the DHT is +/- .2 degrees

Changed the DHT pins used back to the original commented code. The existing pins are the same used for the serial display to the read the values in the Arduino IDE monitor. Changing them allow the serial output to work again.